fix(ruleset): misc changes to rule id's
This commit contains a number of changes
identified in our "final audit" before the 1.0
release is delivered.  Mainly rule-id changes
and other minor things.
