+++ packages/ruleset/src/ibm-oas.js
@@ -95,74 +95,71 @@ module.exports = {
     'oas3-unused-component': true,
 
     // IBM Custom Rules
-    'ibm-accept-header': ibmRules.acceptHeader,
     'ibm-array-attributes': ibmRules.arrayAttributes,
-    'ibm-array-of-arrays': ibmRules.arrayOfArrays,
-    'ibm-array-responses': ibmRules.arrayResponses,
-    'ibm-authorization-header': ibmRules.authorizationHeader,
+    'ibm-avoid-inline-schemas': ibmRules.inlineSchemas,
+    'ibm-avoid-property-name-collision': ibmRules.propertyNameCollision,
+    'ibm-avoid-repeating-path-parameters': ibmRules.duplicatePathParameter,
     'ibm-binary-schemas': ibmRules.binarySchemas,
-    'ibm-circular-refs': ibmRules.circularRefs,
     'ibm-collection-array-property': ibmRules.collectionArrayProperty,
-    'ibm-consecutive-path-segments': ibmRules.consecutivePathSegments,
     'ibm-content-contains-schema': ibmRules.contentContainsSchema,
-    'ibm-content-exists': ibmRules.contentExists,
     'ibm-content-type-is-specific': ibmRules.contentTypeIsSpecific,
-    'ibm-content-type-header': ibmRules.contentTypeHeader,
-    'ibm-delete-body': ibmRules.deleteBody,
-    'ibm-description-mentions-json': ibmRules.descriptionMentionsJSON,
-    'ibm-discriminator-property-exists': ibmRules.discriminatorPropertyExists,
-    'ibm-duplicate-path-parameter': ibmRules.duplicatePathParameter,
+    'ibm-define-required-properties': ibmRules.requiredPropertyMissing,
+    'ibm-discriminator-property': ibmRules.discriminatorPropertyExists,
+    'ibm-dont-require-merge-patch-properties': ibmRules.mergePatchProperties,
     'ibm-enum-casing-convention': ibmRules.enumCasingConvention,
     'ibm-error-content-type-is-json': ibmRules.errorContentTypeIsJson,
     'ibm-error-response-schemas': ibmRules.errorResponseSchemas,
-    'ibm-etag-header-exists': ibmRules.etagHeaderExists,
-    'ibm-examples-name-contains-space': ibmRules.examplesNameContainsSpace,
-    'ibm-if-modified-since-header': ibmRules.ifModifiedSinceHeader,
-    'ibm-if-unmodified-since-header': ibmRules.ifUnmodifiedSinceHeader,
-    'ibm-inline-property-schema': ibmRules.inlinePropertySchema,
-    'ibm-inline-request-schema': ibmRules.inlineRequestSchema,
-    'ibm-inline-response-schema': ibmRules.inlineResponseSchema,
+    'ibm-etag-header': ibmRules.etagHeaderExists,
     'ibm-major-version-in-path': ibmRules.majorVersionInPath,
-    'ibm-merge-patch-properties': ibmRules.mergePatchProperties,
+    'ibm-no-accept-header': ibmRules.acceptHeader,
+    'ibm-no-array-of-arrays': ibmRules.arrayOfArrays,
+    'ibm-no-array-responses': ibmRules.arrayResponses,
+    'ibm-no-authorization-header': ibmRules.authorizationHeader,
+    'ibm-no-body-for-delete': ibmRules.deleteBody,
+    'ibm-no-circular-refs': ibmRules.circularRefs,
+    'ibm-no-consecutive-path-parameter-segments':
+      ibmRules.consecutivePathSegments,
+    'ibm-no-content-type-header': ibmRules.contentTypeHeader,
+    'ibm-no-crn-path-parameters': ibmRules.pathParameterNotCRN,
+    'ibm-no-default-for-required-parameter': ibmRules.parameterDefault,
+    'ibm-no-duplicate-description-with-ref-sibling':
+      ibmRules.refSiblingDuplicateDescription,
+    'ibm-no-if-modified-since-header': ibmRules.ifModifiedSinceHeader,
+    'ibm-no-if-unmodified-since-header': ibmRules.ifUnmodifiedSinceHeader,
+    'ibm-no-optional-properties-in-required-body': ibmRules.optionalRequestBody,
+    'ibm-no-space-in-example-name': ibmRules.examplesNameContainsSpace,
+    'ibm-openapi-tags-used': ibmRules.unusedTags,
     'ibm-operationid-casing-convention': ibmRules.operationIdCasingConvention,
     'ibm-operationid-naming-convention': ibmRules.operationIdNamingConvention,
-    'ibm-operation-summary-exists': ibmRules.operationSummaryExists,
-    'ibm-optional-requestbody': ibmRules.optionalRequestBody,
+    'ibm-operation-summary': ibmRules.operationSummaryExists,
     'ibm-pagination-style': ibmRules.paginationStyle,
     'ibm-parameter-casing-convention': ibmRules.parameterCasingConvention,
-    'ibm-parameter-default': ibmRules.parameterDefault,
-    'ibm-parameter-description-exists': ibmRules.parameterDescriptionExists,
+    'ibm-parameter-description': ibmRules.parameterDescriptionExists,
     'ibm-parameter-order': ibmRules.parameterOrder,
-    'ibm-parameter-schema-or-content-exists':
-      ibmRules.parameterSchemaOrContentExists,
+    'ibm-parameter-schema-or-content': ibmRules.parameterSchemaOrContentExists,
     'ibm-patch-request-content-type': ibmRules.patchRequestContentType,
-    'ibm-path-parameter-not-crn': ibmRules.pathParameterNotCRN,
     'ibm-path-segment-casing-convention': ibmRules.pathSegmentCasingConvention,
-    'ibm-precondition-header': ibmRules.preconditionHeader,
+    'ibm-precondition-headers': ibmRules.preconditionHeader,
     'ibm-property-attributes': ibmRules.propertyAttributes,
     'ibm-property-casing-convention': ibmRules.propertyCasingConvention,
-    'ibm-property-description-exists': ibmRules.propertyDescriptionExists,
     'ibm-property-consistent-name-and-type':
       ibmRules.propertyConsistentNameAndType,
-    'ibm-property-name-collision': ibmRules.propertyNameCollision,
+    'ibm-property-description': ibmRules.propertyDescriptionExists,
     'ibm-ref-pattern': ibmRules.refPattern,
-    'ibm-ref-sibling-duplicate-description':
-      ibmRules.refSiblingDuplicateDescription,
+    'ibm-request-and-response-content': ibmRules.contentExists,
     'ibm-requestbody-is-object': ibmRules.requestBodyIsObject,
-    'ibm-requestbody-name-exists': ibmRules.requestBodyNameExists,
-    'ibm-required-property-missing': ibmRules.requiredPropertyMissing,
-    'ibm-response-example-exists': ibmRules.responseExampleExists,
+    'ibm-requestbody-name': ibmRules.requestBodyNameExists,
     'ibm-response-status-codes': ibmRules.responseStatusCodes,
-    'ibm-schema-description-exists': ibmRules.schemaDescriptionExists,
-    'ibm-schema-type-exists': ibmRules.schemaTypeExists,
+    'ibm-schema-description': ibmRules.schemaDescriptionExists,
+    'ibm-schema-type': ibmRules.schemaTypeExists,
     'ibm-schema-type-format': ibmRules.schemaTypeFormat,
     'ibm-sdk-operations': ibmRules.ibmSdkOperations,
     'ibm-securityscheme-attributes': ibmRules.securitySchemeAttributes,
     'ibm-securityschemes': ibmRules.securitySchemes,
     'ibm-server-variable-default-value': ibmRules.serverVariableDefaultValue,
     'ibm-string-attributes': ibmRules.stringAttributes,
+    'ibm-success-response-example': ibmRules.responseExampleExists,
     'ibm-summary-sentence-style': ibmRules.summarySentenceStyle,
-    'ibm-unused-tags': ibmRules.unusedTags,
     'ibm-valid-path-segments': ibmRules.validPathSegments,
   },
 };
