--- packages/ruleset/src/rules/description-mentions-json.js
@@ -1,22 +1,0 @@
-/**
- * Copyright 2017 - 2023 IBM Corporation.
- * SPDX-License-Identifier: Apache2.0
- */
-
-const {
-  schemas,
-} = require('@ibm-cloud/openapi-ruleset-utilities/src/collections');
-const { oas3 } = require('@stoplight/spectral-formats');
-const { descriptionMentionsJSON } = require('../functions');
-
-module.exports = {
-  description: 'Schema descriptions should avoid mentioning "JSON"',
-  message: '{{error}}',
-  given: schemas,
-  severity: 'warn',
-  formats: [oas3],
-  resolved: true,
-  then: {
-    function: descriptionMentionsJSON,
-  },
-};
