+++ packages/ruleset/src/rules/index.js
@@ -18,7 +18,6 @@ module.exports = {
   contentTypeHeader: require('./content-type-header'),
   contentTypeIsSpecific: require('./content-type-is-specific'),
   deleteBody: require('./delete-body'),
-  descriptionMentionsJSON: require('./description-mentions-json'),
   discriminatorPropertyExists: require('./discriminator-property-exists'),
   duplicatePathParameter: require('./duplicate-path-parameter'),
   etagHeaderExists: require('./etag-header-exists'),
@@ -29,9 +28,7 @@ module.exports = {
   ibmSdkOperations: require('./ibm-sdk-operations'),
   ifModifiedSinceHeader: require('./if-modified-since-header'),
   ifUnmodifiedSinceHeader: require('./if-unmodified-since-header'),
-  inlinePropertySchema: require('./inline-property-schema'),
-  inlineRequestSchema: require('./inline-request-schema'),
-  inlineResponseSchema: require('./inline-response-schema'),
+  inlineSchemas: require('./inline-schemas'),
   majorVersionInPath: require('./major-version-in-path'),
   mergePatchProperties: require('./merge-patch-properties'),
   operationIdCasingConvention: require('./operationid-casing-convention'),
