+++ packages/ruleset/src/rules/inline-schemas.js
@@ -7,7 +7,7 @@ const {
   unresolvedSchemas,
 } = require('@ibm-cloud/openapi-ruleset-utilities/src/collections');
 const { oas3 } = require('@stoplight/spectral-formats');
-const { inlinePropertySchema } = require('../functions');
+const { inlineSchemas } = require('../functions');
 
 module.exports = {
   description: 'Nested objects should be defined as a $ref to a named schema',
@@ -17,6 +17,6 @@ module.exports = {
   severity: 'warn',
   resolved: false,
   then: {
-    function: inlinePropertySchema,
+    function: inlineSchemas,
   },
 };
