+++ packages/ruleset/test/array-responses.test.js
@@ -7,7 +7,7 @@ const { arrayResponses } = require('../src/rules');
 const { makeCopy, rootDocument, testRule, severityCodes } = require('./utils');
 
 const rule = arrayResponses;
-const ruleId = 'ibm-array-responses';
+const ruleId = 'ibm-no-array-responses';
 const expectedSeverity = severityCodes.error;
 const expectedMsg =
   'Operations should not return an array as the top-level structure of a response.';
