+++ packages/ruleset/test/delete-body.test.js
@@ -7,7 +7,7 @@ const { deleteBody } = require('../src/rules');
 const { makeCopy, rootDocument, testRule, severityCodes } = require('./utils');
 
 const rule = deleteBody;
-const ruleId = 'ibm-delete-body';
+const ruleId = 'ibm-no-body-for-delete';
 const expectedSeverity = severityCodes.warning;
 const expectedMsg = '"delete" operation should not contain a requestBody.';
