+++ packages/ruleset/test/duplicate-path-parameter.test.js
@@ -7,7 +7,7 @@ const { duplicatePathParameter } = require('../src/rules');
 const { makeCopy, rootDocument, testRule, severityCodes } = require('./utils');
 
 const rule = duplicatePathParameter;
-const ruleId = 'ibm-duplicate-path-parameter';
+const ruleId = 'ibm-avoid-repeating-path-parameters';
 const expectedSeverity = severityCodes.warning;
 const expectedMsg =
   'Common path parameters should be defined on the path object.';
