+++ packages/ruleset/test/parameter-default.test.js
@@ -7,7 +7,7 @@ const { parameterDefault } = require('../src/rules');
 const { makeCopy, rootDocument, testRule, severityCodes } = require('./utils');
 
 const rule = parameterDefault;
-const ruleId = 'ibm-parameter-default';
+const ruleId = 'ibm-no-default-for-required-parameter';
 const expectedSeverity = severityCodes.warning;
 const expectedMsg = 'Required parameter should not define a default value';
